Move zfs keys related logic in another partial plugin
